Бизнес-логика получения информации о пользователе + тесты

+++ Weight Journal/Authenticate/Model/UserModel.swift
@@ -14,13 +14,13 @@ struct UserInfo: Codable {
     var uid: String
     var caloriesGoal: Int
     var age: Int
-    var height: Int
-    var weightNow: Int
-    var weightGoal: Int
+    var height: Double
+    var weightNow: Double
+    var weightGoal: Double
     var sex: String
     var caloriesNow: Int
-    var max: Int
-    var min: Int
+    var max: Double
+    var min: Double
     var savedFood: [String]
     var foodDate: [String: String]
     var weightDate: [String: String]
